Added proper redirecting upon signin so that if the user has any boards, she will go to /boards to see the list of boards, if the user does not have any boards, a new board will be created and added to that user's boards, and she will be redirected there.
Verified that all of the above is working.

Now working on getting the boards page to render properly (it's not pulling in the data currently), followed by adding proper navigation to move between specific board and /boards, as well as creating a new board from /boards. Final step will be to ensure auth checks are present on all necessary routes